#48 & #61 Made Sign in Possible
Added the files necessary to be able to sign in, sign out, with errors for wrong username, wrong password, now able to ask questions if signed in, the logo now goes to the home page, and the sign in goes to the sign in, everything connected. Hopefully i didnt forget any.